ue size_t, avoid warnings
[[Imported from SVN: r250]]

+++ glue/extractors/cubesurfaceextractor.hh
@@ -431,7 +431,7 @@ void CubeSurfaceExtractor<GV, rect>::globalCoords(unsigned int index, const Coor
   Dune::array<Coords, simplex_corners> corners;
   for (int i = 0; i < simplex_corners; ++i)
     corners[i] = this->_coords[this->_faces[index].corners[i].idx].coord;
-  for (int i = 0; i < bcoords.size(); ++i)
+  for (size_t i = 0; i < bcoords.size(); ++i)
     interpolateBarycentric<simplex_corners, ctype, Dune::FieldVector<ctype, dimworld> >(corners, bcoords[i], wcoords[i], dimworld);
 }
 
@@ -460,7 +460,7 @@ void CubeSurfaceExtractor<GV, rect>::localCoords(unsigned int index,
       for (int i = 0; i < simplex_corners; ++i)
         corners[i] = cornerLocalInRefElement<ctype, dimworld>(gt, num_in_self, 3-i);
     }
-    for (int i = 0; i < elementCoords.size(); ++i)
+    for (size_t i = 0; i < elementCoords.size(); ++i)
       interpolateBarycentric<dimworld, ctype, Dune::FieldVector<ctype, dimworld> >(corners, referenceToBarycentric(subEntityCoords[i]), elementCoords[i], dimworld);
   }
   else
@@ -488,7 +488,7 @@ void CubeSurfaceExtractor<GV, rect>::localAndGlobalCoords(unsigned int index,
   else
   {
     ElementPtr eptr = this->_elmtInfo.find(this->_faces[index].parent)->second->p;
-    for (int i = 0; i < ecoords.size(); ++i)
+    for (size_t i = 0; i < ecoords.size(); ++i)
       ecoords[i] = eptr->geometry().local(wcoords[i]);
   }
 }
